+++ src/main/Core/BrowserWindow/AppIconFilePathResolver/AppIconFilePathResolver.test.ts
