+++ src/main/Core/BrowserWindow/AppIconFilePathResolver/AppIconFilePathResolver.ts
