+++ src/main/Core/BrowserWindow/AppIconFilePathResolver/index.ts
