+++ src/main/Core/BrowserWindow/Contract/BrowserWindowAppIconFilePathResolver.ts
@@ -1,0 +1,3 @@
+export interface BrowserWindowAppIconFilePathResolver {
+    getAppIconFilePath(): string;
+}
