+++ src/main/Core/BrowserWindow/Contract/BrowserWindowVibrancyProvider.ts
