+++ src/main/Core/BrowserWindow/Contract/index.ts
@@ -1,4 +1,5 @@
 export * from "./BackgroundMaterialProvider";
-export * from "./HtmlLoader";
+export * from "./BrowserWindowAppIconFilePathResolver";
+export * from "./BrowserWindowHtmlLoader";
+export * from "./BrowserWindowVibrancyProvider";
 export * from "./Vibrancy";
-export * from "./VibrancyProvider";
