+++ src/main/Core/SearchWindow/BrowserWindowConstructorOptionsProvider/index.ts
@@ -1,5 +1,3 @@
-export * from "../../BrowserWindow/BackgroundMaterial";
-export * from "../../BrowserWindow/Vibrancy";
 export * from "./BrowserWindowConstructorOptionsProvider";
 export * from "./DefaultBrowserWindowConstructorOptionsProvider";
 export * from "./defaultWindowSize";
